Init navigation state completely from history state, refactored view width calc into own function. Prevent unnecessary navigation dispatch on reloading window.
